refactor: reorganize PTY shell package
