+++ src/aish/shell/__init__.py
@@ -1,0 +1,37 @@
+"""Shell runtime package."""
+
+from .entry import run_shell
+from .pty.executor import execute_command_with_pty
+from .runtime.ai import AIHandler
+from .runtime.app import PTYAIShell
+from .runtime.events import LLMEventRouter
+from .runtime.output import OutputProcessor
+from .runtime.router import InputRouter
+from .types import (
+    ActionContext,
+    ActionOutcome,
+    CommandResult,
+    CommandStatus,
+    InputIntent,
+)
+from .ui.interaction import PTYUserInteraction
+from .ui.placeholder import PlaceholderManager
+from .ui.suggestions import SuggestionEngine
+
+__all__ = [
+    "AIHandler",
+    "ActionContext",
+    "ActionOutcome",
+    "CommandResult",
+    "CommandStatus",
+    "InputIntent",
+    "InputRouter",
+    "LLMEventRouter",
+    "OutputProcessor",
+    "PTYUserInteraction",
+    "PTYAIShell",
+    "PlaceholderManager",
+    "SuggestionEngine",
+    "execute_command_with_pty",
+    "run_shell",
+]
