+++ src/aish/shell/pty/__init__.py
@@ -1,0 +1,5 @@
+"""PTY execution helpers for the shell package."""
+
+from .executor import execute_command_with_pty
+
+__all__ = ["execute_command_with_pty"]
