+++ src/aish/shell/runtime/__init__.py
@@ -1,0 +1,15 @@
+"""Runtime-layer shell components."""
+
+from .ai import AIHandler
+from .app import PTYAIShell
+from .events import LLMEventRouter
+from .output import OutputProcessor
+from .router import InputRouter
+
+__all__ = [
+    "AIHandler",
+    "InputRouter",
+    "LLMEventRouter",
+    "OutputProcessor",
+    "PTYAIShell",
+]
