+++ src/aish/shell/runtime/events.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import Callable
 
-from ..llm import LLMCallbackResult, LLMEvent, LLMEventType
+from ...llm import LLMCallbackResult, LLMEvent, LLMEventType
 
 
 @dataclass
