--- src/aish/shell_pty_ui/__init__.py
@@ -1,6 +1,0 @@
-"""PTY-specific user interaction utilities."""
-
-from .interaction import PTYUserInteraction
-from .placeholder_manager import PlaceholderManager
-
-__all__ = ["PTYUserInteraction", "PlaceholderManager"]
